add to other header items
