+++ src/sentry/static/sentry/app/components/organizations/multipleProjectSelector.jsx
@@ -1,9 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import styled, {css} from 'react-emotion';
-import Button from 'app/components/button';
-import {growIn} from 'app/styles/animations';
-import space from 'app/styles/space';
 
 import SentryTypes from 'app/sentryTypes';
 import {analytics} from 'app/utils/analytics';
@@ -13,6 +10,7 @@ import ProjectSelector from 'app/components/projectSelector';
 import InlineSvg from 'app/components/inlineSvg';
 
 import HeaderItem from 'app/components/organizations/headerItem';
+import MultipleSelectorSubmitRow from 'app/components/organizations/multipleSelectorSubmitRow';
 
 const rootContainerStyles = css`
   display: flex;
@@ -132,7 +130,6 @@ export default class MultipleProjectSelector extends React.PureComponent {
   render() {
     const {value, projects, multi, forceProject} = this.props;
     const selectedProjectIds = new Set(value);
-    const showSubmitButton = multi && this.state.hasChanges;
 
     const selected = projects.filter(project =>
       selectedProjectIds.has(parseInt(project.id, 10))
@@ -157,16 +154,8 @@ export default class MultipleProjectSelector extends React.PureComponent {
         onMultiSelect={this.handleMultiSelect}
         rootClassName={rootContainerStyles}
         menuFooter={({actions}) =>
-          showSubmitButton && (
-            <SubmitButtonContainer>
-              <SubmitButton
-                onClick={() => this.handleUpdate(actions)}
-                size="xsmall"
-                priority="primary"
-              >
-                {t('Apply')}
-              </SubmitButton>
-            </SubmitButtonContainer>
+          this.state.hasChanges && (
+            <MultipleSelectorSubmitRow onSubmit={() => this.handleUpdate(actions)} />
           )
         }
       >
@@ -207,16 +196,6 @@ const StyledProjectSelector = styled(ProjectSelector)`
   margin: 1px 0 0 -1px;
   border-radius: 0 0 4px 4px;
   width: 110%;
-`;
-
-const SubmitButtonContainer = styled('div')`
-  display: flex;
-  justify-content: flex-end;
-`;
-
-const SubmitButton = styled(Button)`
-  animation: 0.1s ${growIn} ease-in;
-  margin: ${space(0.5)} 0;
 `;
 
 const StyledHeaderItem = styled(HeaderItem)`
